drivers/uktty: Move late init into MMIO module
This commit moves the FDT-related late init code from the
main NS16550 driver into the MMIO module and updates the
code slightly.
